fix(BackButton navigation): fix redirecting on game bord instead of home when player click back on exit.
When player end a game party and click on exit game, he click again on back button to go back to home  but instead of go back home, the button navigate to the previous game board. But fix it in this commit. Now, the player should be correctly redicted to the home page.